pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shinka"
version = "0.0.1"
description = "Automated Scientific Evolution"
authors = [
    {name = "Sakana AI", email = "robert@sakana.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai",
    "litellm",
    "numpy",
    "pandas",
    "anthropic",
    "requests",
    "boto3",
    "pydantic",
    "backoff",
    "python-dotenv",
    "instructor",
    "python-Levenshtein",
    "radon",
    "unidiff",
    "dill",
    "hydra-core==1.3.2",
    "matplotlib",
    "networkx",
    "seaborn",
    "moviepy",
    "scikit-learn",
    "adjustText",
    "markdown",
    "aiofiles",
    "google-generativeai",
]

[tool.setuptools]
script-files = ["shinka/shinka_launch", "shinka/shinka_visualize"]

[tool.setuptools.packages.find]
include = ["shinka", "shinka.*"]

[tool.setuptools.package-data]
"*" = ["*"]

[dependency-groups]
dev = [
    "pytest>=6.0",
    "black",
    "isort",
    "flake8",
]
